moved Partitioning from ECS to root
